backend/app.py
import logging
import os
import time
import traceback
import uuid

import cv2
import numpy as np
import tensorflow as tf
from dotenv import load_dotenv
from flask import Flask, jsonify, request, send_from_directory
from flask_cors import CORS
from flask_limiter import Limiter
from flask_limiter.util import get_remote_address
from gemini_service import generate_explanation
from gradcam import make_gradcam_heatmap, overlay_heatmap, preprocess_image
from werkzeug.exceptions import HTTPException

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

load_dotenv()

app = Flask(__name__)
CORS(app, resources={r"/*": {"origins": "*"}})

DEFAULT_RATE_LIMIT = os.getenv("DEFAULT_RATE_LIMIT", "60 per minute")
ANALYZE_RATE_LIMIT = os.getenv("ANALYZE_RATE_LIMIT", "20 per minute")
API_KEY = os.getenv("API_KEY", "").strip()

limiter = Limiter(
    key_func=get_remote_address,
    default_limits=[DEFAULT_RATE_LIMIT],
    storage_uri="memory://",
)
limiter.init_app(app)

UPLOAD_FOLDER = "uploads"
OUTPUT_FOLDER = "outputs"

os.makedirs(UPLOAD_FOLDER, exist_ok=True)
os.makedirs(OUTPUT_FOLDER, exist_ok=True)

MODEL_PATH = "model.h5"
model = None

try:
    model = tf.keras.models.load_model(MODEL_PATH)
    logger.info("Model loaded in Flask app from '%s'", MODEL_PATH)
except FileNotFoundError:
    logger.warning(
        "Model file '%s' not found. The /analyze endpoint will fall back to a heuristic "
        "heatmap until the model is provided.",
        MODEL_PATH,
    )
except Exception as e:
    logger.error("Failed to load model from '%s': %s", MODEL_PATH, e, exc_info=True)


def _is_request_authorized(req):
    """
    Enforce API key auth when API_KEY is configured.
    If API_KEY is empty, authentication is disabled for local development.
    """
    if not API_KEY:
        return True
    incoming = req.headers.get("X-API-Key", "").strip()
    return incoming == API_KEY


@app.errorhandler(Exception)
def handle_unexpected_error(error):
    """
    Global JSON error handler so the frontend never receives HTML error pages.
    Logs full traceback and returns a structured JSON error.
    """
    if isinstance(error, HTTPException):
        response = {
            "error": error.name,
            "message": error.description,
            "status_code": error.code,
        }
        logger.error("HTTP %s error: %s", error.code, error, exc_info=True)
        return jsonify(response), error.code

    logger.error("Unhandled exception in request: %s", error, exc_info=True)
    return (
        jsonify(
            {
                "error": "Internal Server Error",
                "message": str(error),
            }
        ),
        500,
    )


@app.route("/health", methods=["GET"])
def health_check():
    return jsonify({"status": "ok"}), 200


@app.route("/analyze", methods=["POST"])
@limiter.limit(ANALYZE_RATE_LIMIT)
def analyze():

    if not _is_request_authorized(request):
        return jsonify({"error": "Unauthorized", "message": "Missing or invalid API key."}), 401

    if "image" not in request.files:
        return jsonify({"error": "No image uploaded"}), 400

    file = request.files["image"]

    # Always persist the originally uploaded image at full resolution.
    # The model will see a resized version, but the saved file is untouched.
    filename = str(uuid.uuid4()) + ".png"
    filepath = os.path.join(UPLOAD_FOLDER, filename)
    file.save(filepath)

    # Measure end-to-end processing time
    started_at = time.perf_counter_ns()

    if model is None:
        # Fallback path when the trained model file is not available.
        # This keeps the API and UI fully functional with deterministic,
        # image-dependent values instead of a constant output.
        original_bgr = cv2.imread(filepath)
        if original_bgr is None:
            return jsonify({"error": "Uploaded image could not be read by OpenCV."}), 400

        original = cv2.cvtColor(original_bgr, cv2.COLOR_BGR2RGB)

        # Create a pseudo-heatmap from a lightly blurred grayscale image.
        gray = cv2.cvtColor(original, cv2.COLOR_RGB2GRAY)
        gray_blur = cv2.GaussianBlur(gray, (15, 15), 0)
        norm_heatmap = cv2.normalize(gray_blur.astype(np.float32), None, 0.0, 1.0, cv2.NORM_MINMAX)

        result = overlay_heatmap(norm_heatmap, original)

        # Save heatmap overlay as PNG to avoid additional JPEG compression loss.
        output_filename = str(uuid.uuid4()) + ".png"
        output_path = os.path.join(OUTPUT_FOLDER, output_filename)
        cv2.imwrite(output_path, cv2.cvtColor(result, cv2.COLOR_RGB2BGR))

        activation_strength = float(norm_heatmap.mean())

        # Simple heuristic so that probabilities vary across images.
        mean_intensity = float(gray.mean() / 255.0)
        real_probability = max(0.05, min(0.95, mean_intensity))
        fake_probability = 1.0 - real_probability

        label = "REAL" if real_probability >= fake_probability else "FAKE"
        confidence = float(max(real_probability, fake_probability) * 100)
    else:
        # Full inference path using the trained model and Grad-CAM.
        img_array, original = preprocess_image(filepath)

        # Robustly extract a scalar prediction regardless of output shape.
        prediction_raw = model.predict(img_array)
        pred = float(np.ravel(prediction_raw)[0])

        real_probability = pred
        fake_probability = 1.0 - pred

        if pred > 0.5:
            label = "REAL"
            confidence = float(pred * 100)
        else:
            label = "FAKE"
            confidence = float((1 - pred) * 100)

        heatmap = make_gradcam_heatmap(img_array, model)
        result = overlay_heatmap(heatmap, original)

        # Save heatmap overlay as PNG at the original image resolution.
        output_filename = str(uuid.uuid4()) + ".png"
        output_path = os.path.join(OUTPUT_FOLDER, output_filename)
        cv2.imwrite(output_path, cv2.cvtColor(result, cv2.COLOR_RGB2BGR))

        activation_strength = float(heatmap.mean())

    finished_at = time.perf_counter_ns()
    inference_time_ms = int((finished_at - started_at) / 1_000_000)

    explanation = generate_explanation(
        user_question="Why is this image classified this way?",
        label=label,
        confidence=round(confidence, 2),
        activation_strength=round(activation_strength, 3),
    )

    return jsonify(
        {
            "label": label,
            "confidence": round(confidence, 2),
            "real_probability": round(real_probability, 4),
            "fake_probability": round(fake_probability, 4),
            "inference_time_ms": inference_time_ms,
            "activation_strength": round(activation_strength, 4),
            "model_version": "cnn-mobilenetv2-v1",
            "heatmap_url": f"/outputs/{output_filename}",
            "explanation": explanation,
        }
    )


@app.route("/outputs/<filename>")
def get_output(filename):
    return send_from_directory(OUTPUT_FOLDER, filename)


# ==============================

if __name__ == "__main__":
    try:
        logger.info("Starting Flask server on http://127.0.0.1:5000")
        app.run(host="127.0.0.1", port=5000, debug=True)
    except Exception:
        # If the server crashes on startup, print full traceback to the console.
        traceback.print_exc()
