backend/requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
Flask-Limiter>=3.10.0
python-dotenv>=1.0.0
tensorflow>=2.15.0
opencv-python>=4.8.0
numpy>=1.24.0,<2.0.0
google-genai>=1.0.0
